Use OAuth2 package for authentication
Adding in logging makes development and debugging a bit easier. Note
that the client UUID and client secret can be leaked in plain text in
error messages and stacktraces in stdout. I haven't figured out the
optimal solution for handling secrets safely yet.

Ideally I don't want the token to be retrieved in the GUI initialization
`SharesightDesktop#init/1`. I tried a memoization strategy for
initializing and getting the access token similar to what is commonly
done for Ruby instance variables. But that got quite confusing fast.
I might try an Agent for storing the access token in the future.

For now at this stage the program can make get requests to the API which
will always fail authentication. The access token can be retrieved it
just needs to be added to the request headers next.