Fetch client environment variables and log a warning on error
This morning the OAuth client was having errors when trying to retrieve
the access token. It turned out the CLIENT_UID environment variable was
misspelled. Logging a warning made it easier to spot the error during
development.

For the future I'd like the OAuth client to not attempt to retrieve the
access token if CLIENT_UID or CLIENT_SECRET are empty strings or nil.