API call is automatic so url text, button, and body text are not needed
Since the app can now display the performance report in the table the
other widgets are not needed anymore. The GET request happens at the end
of the GUI initialization. The performance report url is fetched from an
environment variable. I'd like to move this and other configuration to a
external configuration file in the future.

One issue that has happened several times is a time out when the GET
request in `ApiClient` is made on line 19. Some error handling needs to
be built in. Currently the app closes with an error. Ideally the app
should be able to fully load while the GET request is handled
asynchronously. If the request fails then the user should be able to
resend it without having to restart the application.