Handle an empty/nil API_URL and provide for the user to reconnect to the API
I've added a `handle_event` callback for the 'Refresh' menu item. At the
moment the callback is too broad as it would catch any command_menu_selected
event from the menu without knowing which menu item was clicked. For now
it doesn't matter since there is only one menu item. There might be a
index value passed with the event that can be used to tell which menu
item was clicked.

wxID_ANY is a build in ID value. It's used in cases where we don't care
about the ID.